requirements.txt
python-telegram-bot[job-queue]==20.7
python-dotenv==1.0.0
asyncpg==0.29.0
aiohttp==3.9.1
aiohttp-cors==0.7.0
Pillow==10.1.0
requests==2.31.0
faster-whisper==1.0.3
pydub==0.25.1
openai>=1.0.0
tiktoken>=0.5.0
boto3>=1.34.0
openpyxl>=3.1.0
reportlab>=4.0
pypdf>=4.0
python-docx>=1.1.0
PyMuPDF>=1.24.0
